Separate example func and add README.md
